refactoring pagerank computation for reducing floating point divisions and changing floats to doubles for pagerank computation.

+++ application/page_rank/edge_data.h
@@ -49,7 +49,7 @@ namespace nimbus {
 struct EdgeEntry {
   size_t src_id;
   size_t dst_id;
-  float delta;
+  double delta;
 };
 
 class EdgeData : public Data {
